requirements.txt
couchbase==4.1.1
python-dotenv==0.21.1
